First solution to item loading problem
Can now convert material to craftable (aka if manager first loads an
item as a material, but later learns it is a craftable instead).

Unintended but good: can save prices of items as if they were materials
and add in recipes at a later time.

Added Auto-Craft[II] recipe tree.